App: don't obliterate app.==
Mongoid defines it's own ==() and eql?() methods that are based on the
ObjectId. Comparable provides it's own ==() function that uses `<=>`
under the hood.  == is used when inserting into a Hash, App docuemnts
are put in hashes hundreds of times per request. We don't was to to call
<=> hundreds of time.

Putting the Comparable mixin at the top allows the mongoid mixin to
override == with it's faster implementation.