Spec: fix failing tests cause of Identity map
Introduction of the identity map caused some failing specs:
https://travis-ci.org/errbit/errbit/jobs/5383844

Though it seems to be less of an issue in recent runs, not sure why.
The identity map is being enabled in the testing thread so the following
test breaks:

 1. get record # added to identity map
 2. call delete api # (doesn't affect identity map of this process)
 3. get record # still in identity map (and thus the test fails)

It was tricky disabling the Identity map in the correct context (where
the specs actually run).