scope the type selector to the choose section
Formerly if any input in the sub-menus had the substring 'type' in its
name, this event handler would have been bound to clicks on the input
element. When this happens, the variable 'chosen' would generally be an
empty string and that would result in an erroneous jQuery finder:
wrapper.find('div.'+section_class+'.'+chosen) causing an unhandled
error.

It looks like the intent is to bind only to the selectable choice icons,
so that's what this commit does and as a bonus, the embedded forms can
now have an input with the word 'type' in its name.