Add date filter for problems/search
Added two input fields with a javascript datepicker to problems/search form to filter problems by date range.

  * The date a problem was first noticed is considered when searching. Example: Problem which was first noticed at `2014-01-05` will be found if your search criteria is `2014-01-01 to 2014-02-01`.

  * The date a problem was last noticed is considered when searching. Example: Problem which was last noticed at `2014-01-05` but first noticed at `2013-12-24` will be found if your criteria is `2014-01-01 to 2014-02-01`.

  * The last day of the date range is fully considered in the search. Example: Problem which was first noticed on `2014-03-09` at `06:00` will be found if your search criteria is `2014-03-01 to 2013-03-09`.

By default the first notice date of the very first problem is used as the beginning value of the date range. The end value is by default the end of today. So searching for something but not a date will not lead to wrong results.