Add date filter for problems/search
Added two input fields with a javascript datepicker to problems/search form to filter problems by date range. first_notice_at and last_notice_at are both considered when searching:

  * Problem which was first noticed at 2014-01-05 will be found if your search criteria is for example 2014-01-01 until 2014-02-01.

  * Problem which was last noticed at 2014-01-05 but first noticed at 2013-12-24 will also be found if your criteria is for example 2014-01-01 until 2014-02-01.

By default the very first notice date of the very first problem is used as the beginning value of the date range. The end value is by default the date of tomorrow. So searching for something but not a date will not lead to wrong results.