Allow users to see all apps
The default behaviour in errbit is that users can 'watch' an
application, which allows them to see it in the web interface and also
means that they receive email notifications of errors.

GDS has a different requirement which is that all users should be able
to see all apps in the web interface and to choose which apps they
'watch' for email notifications.

This commit changes the various controllers and UI to support this other
approach.