Upgrade Rails to 4.1.14.1
To mitigate potential security vulnerabilities:
http://weblog.rubyonrails.org/2016/1/25/Rails-5-0-0-beta1-1-4-2-5-1-4-1-14-1-3-2-22-1-and-rails-html-sanitizer-1-0-3-have-been-released/

I have not checked if the app is vulnerable to these issues.

The `govuk_security_audit` gem reports several other possible issues
with Gems used by this app, which this pull request does not address. We
can address those separately.

I was unable to upgrade simply by changing `RAILS_VERSION` and running
`bundle install`, apparently due to cyclical dependencies. Also,
deleting the `Gemfile.lock` was not an option because many of the gems
are not pinned in the `Gemfile`, meaning I'd inadvertently upgrade
unrelated gems. So I ran:

    sed -i '' 's/4.1.11/4.1.14.1/g' Gemfile.lock

...which allowed `bundle install` to run successfully.