fix tmp directory permissions
ordering of commands is important here because when asset:precompile is run, it's done as root, which then creates the /app/tmp file with root permissions. Subsequently, after the docker image is started, when connecting to the website for the first time you get a Permission denied error from Devise.