Allow sending email notification for *every* error
This makes the behaviour of `ERRBIT_EMAIL_AT_NOTICES` consistent with
that of `ERRBIT_NOTIFY_AT_NOTICES`.

We work hard to keep the exception notifications in our apps to zero.
The default configuration is to only send emails for the 1st, 10th and
100th occurrence of an error. We've found that this sometimes means we
mark the 1st occurrence error as resolved and then it recurs a few more
times and we are not notified. We'd prefer to be notified every time,
hence this change.