Repair our slack format
We do not want to use the large format from the default. A more concise
format is helpful in slack to not drown errors in whitespace.

Repair the slack format back to something similar to our old format.
This is actually an intermediary format, something similar to both our
old, and the new default message format.