Fixes error on long multibyte string messages
Mongo accepts index keys only if they are less than 1,024 bytes
including some number of bytes of BSON overhead as defined by the BSON
spec. I'm not sure if that number of bytes is predictable and the value
of getting every last possible byte of this message in mongo seems low
since we're arbitrarily truncating anyway. So 1,000 bytes it is.