#1385: Can't merge errors after search (#1411)
The issue is related to the fact that the `#problem_table` is emptied
when the search is run, and a new table is inserted in its place with
result. The existing click binding of the buttons at the end of
the table (merge/delete/etc) is removed when the table is removed.

As a fix, `assets/javascripts/errbit.js` introduces global
`bindProblemButtonsActions` function which is extracted from the
body of `errbit.js`, containing handler of for problem buttons click
event. This function is then added to the result of
`views/problems/search.js.haml` to be run on a new problems table
when the search is returned.