:bug: Fix Google authentication due the sunset from Legacy People API
On March/7/2019 Google has officialy turn off the Legacy People API, keeping it alive just for old projects. More infos in https://developers.google.com/people/legacy. For new projects we have to use the new People API, but the current version from omniauth-google-oauth2 on the project is still not compatible and that's completely breaks the Google's authentication flow from the application.

To fix this issue, I'm submiting this change that updates omniauth-google-oauth2 version.